+++ retdec/tools/decompiler.py
@@ -264,14 +264,27 @@ def parse_args(argv):
     )
     _add_arguments_shared_by_all_tools(parser)
     parser.add_argument(
+        '-a', '--architecture',
+        dest='architecture',
+        choices=['x86', 'arm', 'thumb', 'mips', 'pic32', 'powerpc'],
+        help='architecture to force when (de)compiling'
+    )
+    parser.add_argument(
         '-b', '--brief',
         dest='brief',
         action='store_true',
         help='print fewer information during the decompilation'
     )
     parser.add_argument(
+        '-f', '--file-format',
+        dest='file_format',
+        choices=['elf', 'pe'],
+        help='file format to force when compiling'
+    )
+    parser.add_argument(
         '-l', '--target-language',
         dest='target_language',
+        choices=['c', 'py'],
         help='target high-level language'
     )
     parser.add_argument(
@@ -368,6 +381,8 @@ def main(argv=None):
     add_decompilation_param_when_given(args, params, 'pdb_file')
     add_decompilation_param_when_given(args, params, 'mode')
     add_decompilation_param_when_given(args, params, 'target_language')
+    add_decompilation_param_when_given(args, params, 'architecture')
+    add_decompilation_param_when_given(args, params, 'file_format')
     add_decompilation_param_when_given(args, params, 'generate_archive')
     decompilation = decompiler.start_decompilation(**params)